requirements.txt
Django==3.1.5
pillow==8.1.0
django-crispy-forms==1.11.1